ADEN-8986 Move example pages configuration
