+++ examples/templates/abcd/script.ts
@@ -1,5 +1,6 @@
 import { AdEngine, context, templateService } from '@wikia/ad-engine';
 import { BigFancyAdAbove } from '@wikia/ad-products';
+import { getConfig as getBigFancyAdAboveConfig } from '../../big-fancy-ad-above-config';
 import customContext from '../../context';
 import '../../styles.scss';
 
@@ -10,6 +11,6 @@ customContext.templates.bfaa = {
 };
 context.extend(customContext);
 
-templateService.register(BigFancyAdAbove);
+templateService.register(BigFancyAdAbove, getBigFancyAdAboveConfig());
 
 new AdEngine().init();
