+++ examples/templates/hivi-uap-ab/script.ts
@@ -5,6 +5,7 @@ import {
 	FloatingRail,
 	setupNpaContext,
 } from '@wikia/ad-products';
+import { getConfig as getBigFancyAdAboveConfig } from '../../big-fancy-ad-above-config';
 import customContext from '../../context';
 import '../../styles.scss';
 
@@ -19,7 +20,7 @@ if (document.body.offsetWidth < 728) {
 
 setupNpaContext();
 
-templateService.register(BigFancyAdAbove);
+templateService.register(BigFancyAdAbove, getBigFancyAdAboveConfig());
 templateService.register(BigFancyAdBelow);
 templateService.register(FloatingRail);
