+++ examples/templates/hivi-uap-static/script.ts
@@ -1,5 +1,6 @@
 import { AdEngine, context, templateService } from '@wikia/ad-engine';
 import { BigFancyAdAbove, BigFancyAdBelow, FloatingRail } from '@wikia/ad-products';
+import { getConfig as getBigFancyAdAboveConfig } from '../../big-fancy-ad-above-config';
 import customContext from '../../context';
 import '../../styles.scss';
 
@@ -12,7 +13,7 @@ if (document.body.offsetWidth < 728) {
 	context.set('targeting.skin', 'fandom_mobile');
 }
 
-templateService.register(BigFancyAdAbove);
+templateService.register(BigFancyAdAbove, getBigFancyAdAboveConfig());
 templateService.register(BigFancyAdBelow);
 templateService.register(FloatingRail);
