+++ examples/templates/hivi-uap/script.ts
@@ -11,6 +11,7 @@ import {
 	setupNpaContext,
 } from '@wikia/ad-products';
 import { AdInfoContext, slotTracker, slotTrackingMiddleware } from '@wikia/ad-tracking';
+import { getConfig as getBigFancyAdAboveConfig } from '../../big-fancy-ad-above-config';
 import customContext from '../../context';
 import '../../styles.scss';
 
@@ -25,7 +26,7 @@ if (document.body.offsetWidth < 728) {
 
 setupNpaContext();
 
-templateService.register(BigFancyAdAbove);
+templateService.register(BigFancyAdAbove, getBigFancyAdAboveConfig());
 templateService.register(BigFancyAdBelow);
 templateService.register(FloatingRail);
